move metabot to use plugins
