+++ enterprise/frontend/src/embedding-sdk-ee/metabot/MetabotChatHistory.tsx
