+++ enterprise/frontend/src/embedding-sdk-ee/metabot/MetabotChatInput.tsx
