+++ enterprise/frontend/src/embedding-sdk-ee/metabot/MetabotQuestionEmptyState.tsx
