+++ enterprise/frontend/src/embedding-sdk-ee/metabot/QuestionDetails.tsx
