+++ enterprise/frontend/src/embedding-sdk-ee/metabot/QuestionTitle.tsx
