+++ enterprise/frontend/src/embedding-sdk-ee/metabot/SidebarHeader.tsx
