+++ enterprise/frontend/src/embedding-sdk-ee/metabot/index.ts
@@ -1,0 +1,5 @@
+// Contains side effects that activate the plugin
+import "./MetabotQuestion";
+
+// the types are in the OSS code, but we're re-exporting them here for convenience to not have to import from two places
+export type { MetabotQuestionProps } from "embedding-sdk-bundle/components/public/MetabotQuestion/types";
