+++ enterprise/frontend/src/metabase-enterprise/sdk-plugins.ts
@@ -27,3 +27,4 @@ export function initializePlugins() {
 // "SDK EE-plugins", that are specific to the embedding sdk.
 // These only apply to the SDK, not to the core app
 import "../embedding-sdk-ee/auth";
+import "../embedding-sdk-ee/metabot";
